images/replication-job/requirements.txt
mapcoloniesjsonlogger==0.0.14
osmeterium==0.0.1
